Add Premake build

+++ scripts/distrib.sh
@@ -2,11 +2,13 @@
 
 ROOT="$(realpath $(dirname $0)/..)"
 
+if [ $# -lt 4 ]; then exit; fi
+
 rm -r "$1"
 rm "$4.zip"
 mkdir -p "$1"
 $ROOT/scripts/copyfiles.sh "$1" "$2" "$3"
-cp "$ROOT/build/unix/$2/$3/Game.$2" "$1"
+cp "$ROOT/build/$2/$3/Game.$2" "$1"
 cp -r "$ROOT/deps-public/licenses" "$1"
 cp "$ROOT/LICENSE.txt" "$1/licenses"
 OUTDIR=$(pwd)
